Restructure files to split styles and components
* Create two folders to handle components and styles
* Rewrite paths to meet restructured format
