+++ src/component/app.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import "./app.scss";
+import "./../style/app.scss";
 
 const App = () => {
     return (
