+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from 'react-dom';
 
-import App from './app';
+import App from './component/app.jsx';
 
 render(<App />, document.getElementById('app'));
